Profile management added, column admin added in the db

+++ app.py
@@ -54,7 +54,7 @@ def create_user():
     password = data.get("password")
 
     if username and password:
-        user = User(username=username, password=password)
+        user = User(username=username, password=password, role='user')
         db.session.add(user)
         db.session.commit()
         return jsonify({"Message": "User Created!!"})
@@ -78,6 +78,9 @@ def update_user(id_user):
     user = User.query.get(id_user)
     data = request.json
 
+    if id_user != current_user.id and current_user.role == "user":
+        return jsonify({"Message":"Operation not allowed"}), 403
+
     if user and data.get("password"):
         user.password = data.get("password")
         db.session.commit()
@@ -93,6 +96,9 @@ def update_user(id_user):
 def delete_user(id_user):
     user = User.query.get(id_user)
 
+    if current_user.role != 'admin':
+        return jsonify({"Message":"Operation not allowed"}), 403
+    
     if user and id_user != current_user.id:
         db.session.delete(user)
         db.commit()
